requirements-gpu.txt
tensorflow-gpu==1.4.0
numpy==1.14.3
pandas==0.22.0
Keras==2.1.1
matplotlib==2.2.2
pydot-ng==1.0.0
networkx==2.1
sklearn==0.0
h5py==2.8.0
nltk==3.3
